refactor: split block_basic files for Q-DUPLICATION-04
